models/prune_and_quantize/requirements.txt
torch==1.9.1  
transformers==4.10.2  
psutil==5.8.0
google-cloud-storage==1.42.0
google-auth==1.35.0
matplotlib==3.4.3
